pyproject.toml
[tool.poetry]
name = "task-tracker"
version = "0.1.0"
description = "Simple task tracker"
authors = ["Taras Dorosh <tdorosh4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlmodel = "^0.0.22"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
bcrypt = "4.0.1"
ruff = "^0.7.0"
pydantic-settings = "^2.6.0"
alembic = "^1.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
